chore: Move components to App directory and update import paths.
