+++ app/dummy/lectureData.ts
