+++ app/dummy/quotesData.ts
