+++ app/types/course/categories.ts
