+++ app/types/course/courseModel.ts
