+++ app/types/course/courseType.ts
