+++ app/types/knowledge/lecture.ts
