+++ app/types/lectureFrom.ts
